Apprication details changes done
