+++ src/App.tsx
@@ -14,6 +14,8 @@ import colors from './app/constant/colors';
 import VersionContext from './app/context/version.context';
 import {CheckVersionResponse} from 'react-native-check-version';
 
+import { GestureHandlerRootView } from 'react-native-gesture-handler';
+
 const queryClient = new QueryClient();
 
 const App = () => {
@@ -21,7 +23,7 @@ const App = () => {
   const versionContextValue = useState<CheckVersionResponse | null>(null);
 
   return (
-    <>
+    <GestureHandlerRootView style={{ flex: 1 }}>
       <VersionContext.Provider value={versionContextValue}>
         <UserContext.Provider value={userContextValue}>
           <Interceptor>
@@ -36,7 +38,7 @@ const App = () => {
         </UserContext.Provider>
       </VersionContext.Provider>
       <Toast />
-    </>
+    </GestureHandlerRootView>
   );
 };
