+++ src/app/Peerly/components/RatingBar.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {View, Text, StyleSheet, Pressable} from 'react-native';
 import Slider from '@react-native-community/slider';
-import {RatingStarIcon} from '../constants/icons';
+import {GoldenStar} from '../constants/icons';
 
 interface RatingBarProps {
   reward: number;
@@ -19,7 +19,7 @@ const RatingBar: React.FC<RatingBarProps> = ({reward, setReward, disabled}) => {
         step={1}
         value={reward}
         onSlidingComplete={value => setReward(value)}
-        thumbImage={RatingStarIcon}
+        thumbImage={GoldenStar}
         minimumTrackTintColor="#FFD700"
         maximumTrackTintColor="#DDD"
         disabled={disabled || false}
@@ -58,23 +58,26 @@ const styles = StyleSheet.create({
   labelsContainer: {
     flexDirection: 'row',
     justifyContent: 'space-between',
-    width: '85%',
+    width: '90%',
   },
   sliderGood: {
     fontSize: 14,
     color: 'black',
     marginLeft: 60,
+    fontWeight : 'bold'
   },
   sliderNice: {
     fontSize: 14,
     color: 'black',
     paddingLeft: 25,
+    fontWeight : 'bold'
   },
   sliderLove: {
     fontSize: 14,
     color: 'black',
     paddingLeft: 10,
     paddingRight: 10,
+    fontWeight : 'bold'
   },
 });
