+++ src/app/Peerly/components/RewardAcknowledgementModal.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {Modal, StyleSheet, Text, View} from 'react-native';
+import { Modal, StyleSheet, Text, View } from 'react-native';
 import Button from './button/button';
 
 interface RewardAcknowledgementModalProps {
@@ -9,6 +9,7 @@ interface RewardAcknowledgementModalProps {
   resetModal: () => void;
   handleConfirm: () => void;
 }
+
 const RewardAcknowledgementModal: React.FC<RewardAcknowledgementModalProps> = ({
   visible,
   isLoading,
@@ -20,19 +21,17 @@ const RewardAcknowledgementModal: React.FC<RewardAcknowledgementModalProps> = ({
     <Modal transparent={true} visible={visible} onRequestClose={resetModal}>
       <View style={styles.modalOverlay}>
         <View style={styles.modalContent}>
-          <Text style={styles.text}>You have given {rewardLabel} reward.</Text>
+          <Text style={styles.text}>
+            You have given <Text style={styles.highlight}>{rewardLabel}</Text> reward.
+          </Text>
           <View style={styles.btnWrapper}>
-            <View style={styles.btnSecondary}>
-              <Button title="reset" type="secondary" onPress={resetModal} />
-            </View>
-            <View>
-              <Button
-                title="Confirm"
-                type="primary"
-                isLoading={isLoading}
-                onPress={handleConfirm}
-              />
-            </View>
+            <Button title="Reset" type="secondary" onPress={resetModal} />
+            <Button
+              title="Confirm"
+              type="primary"
+              isLoading={isLoading}
+              onPress={handleConfirm}
+            />
           </View>
         </View>
       </View>
@@ -49,23 +48,30 @@ const styles = StyleSheet.create({
   },
   modalContent: {
     backgroundColor: 'white',
-    padding: 20,
-    borderRadius: 10,
+    padding: 30,
+    borderRadius: 20,
     width: '80%',
     alignItems: 'center',
-    justifyContent: 'center',
+    elevation: 10,
+    shadowColor: '#000',
+    shadowOpacity: 0.25,
+    shadowRadius: 4,
+    shadowOffset: { width: 0, height: 2 },
   },
   text: {
-    marginBottom: 15,
+    fontSize: 18,
+    textAlign: 'center',
+    marginBottom: 25,
+    color: '#333',
+  },
+  highlight: {
+    fontWeight: 'bold',
+    color: '#4CAF50',
   },
   btnWrapper: {
     flexDirection: 'row',
     justifyContent: 'space-between',
-    alignItems: 'center',
-  },
-  btnSecondary: {
-    width: '30%',
-    marginRight: 5,
+    width: '100%',
   },
 });
